+++ raster/rt_core/rt_api.h
@@ -500,7 +500,7 @@ double rt_band_get_nodata(rt_band band);
 int rt_band_set_pixel_line(
 	rt_band band,
 	int x, int y,
-	void *vals, uint16_t len
+	void *vals, uint32_t len
 );
 
 /**
